only run full tests on linux

+++ recipe/run_test.sh
@@ -1,18 +1,20 @@
 #!/bin/bash
+if [ "$(uname)" != "Darwin" ]; then
 
-unset REQUESTS_CA_BUNDLE
-unset SSL_CERT_FILE
+    unset REQUESTS_CA_BUNDLE
+    unset SSL_CERT_FILE
 
-export DB_PATH="$SRC_DIR/temp-mongo-db"
-export LOG_PATH="$SRC_DIR/mongolog"
-export DB_PORT=27272
-export PID_FILE_PATH="$SRC_DIR/mongopidfile"
+    export DB_PATH="$SRC_DIR/temp-mongo-db"
+    export LOG_PATH="$SRC_DIR/mongolog"
+    export DB_PORT=27272
+    export PID_FILE_PATH="$SRC_DIR/mongopidfile"
 
-mkdir "$DB_PATH"
+    mkdir "$DB_PATH"
 
-mongod --dbpath="$DB_PATH" --fork --logpath="$LOG_PATH" --port="$DB_PORT" --pidfilepath="$PID_FILE_PATH"
+    mongod --dbpath="$DB_PATH" --fork --logpath="$LOG_PATH" --port="$DB_PORT" --pidfilepath="$PID_FILE_PATH"
 
-python setup.py test
+    python setup.py test
 
-# Terminate the forked process after the test suite exits
-kill `cat $PID_FILE_PATH`
+    # Terminate the forked process after the test suite exits
+    kill `cat $PID_FILE_PATH`
+fi
